chore(tabbar): split demo
