+++ src/packages/__VUE/tabbar/doc.md
@@ -17,340 +17,41 @@ app.use(TabbarItem);
 
 ### 基础用法
 
-:::demo
-
-```vue
-<template>
-  <nut-tabbar @tab-switch="tabSwitch">
-    <nut-tabbar-item tab-title="标签">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签">
-      <template #icon>
-        <Category></Category>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签">
-      <template #icon>
-        <Find></Find>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签">
-      <template #icon>
-        <Cart></Cart>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签">
-      <template #icon>
-        <My></My>
-      </template>
-    </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-import { Home, Category, Find, Cart, My } from '@nutui/icons-vue';
-const tabSwitch = (item, index) => {
-  console.log(item, index);
-};
-</script>
-```
-
-:::
+> demo: tabbar basic
 
 ### 通过名称匹配
 
-:::demo
-
-```vue
-<template>
-  <nut-tabbar v-model="activeName" @tab-switch="tabSwitch">
-    <nut-tabbar-item v-for="(item, index) in List" :name="item.name" :tab-title="item.title" :icon="item.icon">
-    </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-import { h, ref } from 'vue';
-import { Home, Category, Find, Cart, My } from '@nutui/icons-vue';
-const activeName = ref('category');
-const List = [
-  {
-    title: '标签',
-    icon: h(Home),
-    name: 'home'
-  },
-  {
-    title: '标签',
-    icon: h(Category),
-    name: 'category'
-  },
-  {
-    title: '标签',
-    icon: h(Find),
-    name: 'find'
-  },
-  {
-    title: '标签',
-    icon: h(Cart),
-    name: 'cart'
-  },
-  {
-    title: '标签',
-    icon: h(My),
-    name: 'my'
-  }
-];
-const tabSwitch = (item, index) => {
-  console.log(item, index);
-};
-</script>
-```
-
-:::
+> demo: tabbar by-name
 
 ### 自定义图标
 
 通过 icon 插槽自定义图标，可以通过 slot-scope 判断标签是否选中
 
-:::demo
-
-```vue
-<template>
-  <nut-tabbar @tab-switch="tabSwitch">
-    <nut-tabbar-item tab-title="标签">
-      <template #icon="props">
-        <img :src="props.active ? icon.active : icon.unactive" alt="" />
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签">
-      <template #icon="props">
-        <img :src="props.active ? icon.active : icon.unactive" alt="" />
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签">
-      <template #icon="props">
-        <img :src="props.active ? icon.active : icon.unactive" alt="" />
-      </template>
-    </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-const tabSwitch = (item, index) => {
-  console.log(item, index);
-};
-const icon = {
-  active: 'https://img11.360buyimg.com/imagetools/jfs/t1/70423/4/20553/3652/62e74629E23ba550e/aeeed0e3b9f43ae6.png',
-  unactive: 'https://img13.360buyimg.com/imagetools/jfs/t1/23319/19/18329/3084/62e7c346E957c54ef/6c3e8a49e52b76f2.png'
-};
-</script>
-```
-
-:::
+> demo: tabbar custom-icon
 
 ### 自定义选中
 
-:::demo
-
-```vue
-<template>
-  <nut-tabbar v-model="active">
-    <nut-tabbar-item v-for="(item, index) in List" :tab-title="item.title" :icon="item.icon"> </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-import { h, ref } from 'vue';
-import { Home, Category, Find, Cart, My } from '@nutui/icons-vue';
-const active = ref(2);
-const List = [
-  {
-    title: '标签',
-    icon: h(Home),
-    name: 'home'
-  },
-  {
-    title: '标签',
-    icon: h(Category),
-    name: 'category'
-  },
-  {
-    title: '标签',
-    icon: h(Find),
-    name: 'find'
-  },
-  {
-    title: '标签',
-    icon: h(Cart),
-    name: 'cart'
-  },
-  {
-    title: '标签',
-    icon: h(My),
-    name: 'my'
-  }
-];
-const tabSwitch = (item, index) => {
-  console.log(item, index);
-};
-</script>
-```
-
-:::
+> demo: tabbar custom-check
 
 ### 徽标提示
 
 设置 dot 属性后，会在图标右上角展示一个小红点；设置 value 属性后，会在图标右上角展示相应的徽标。
 
 内部使用了 bardge 徽标组件，可将其属性直接作用于 TabbarItem props。
-:::demo
 
-```vue
-<template>
-  <nut-tabbar>
-    <nut-tabbar-item tab-title="标签" :value="11">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签" :dot="true">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签" :value="110">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-import { Home, Category, Find, Cart, My } from '@nutui/icons-vue';
-</script>
-```
-
-:::
+> demo: tabbar show-badge
 
 ### 自定义颜色
 
-:::demo
-
-```vue
-<template>
-  <nut-tabbar unactive-color="#7d7e80" active-color="#1989fa">
-    <nut-tabbar-item v-for="(item, index) in List" :tab-title="item.title" :icon="item.icon"> </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-import { h, ref } from 'vue';
-import { Home, Category, Find, Cart, My } from '@nutui/icons-vue';
-const List = [
-  {
-    title: '标签',
-    icon: h(Home),
-    name: 'home'
-  },
-  {
-    title: '标签',
-    icon: h(Category),
-    name: 'category'
-  },
-  {
-    title: '标签',
-    icon: h(Find),
-    name: 'find'
-  },
-  {
-    title: '标签',
-    icon: h(Cart),
-    name: 'cart'
-  },
-  {
-    title: '标签',
-    icon: h(My),
-    name: 'my'
-  }
-];
-</script>
-```
-
-:::
+> demo: tabbar custom-color
 
 ### 自定义数量
 
-:::demo
-
-```vue
-<template>
-  <nut-tabbar unactive-color="#7d7e80" active-color="#1989fa">
-    <nut-tabbar-item tab-title="标签" :value="11">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签" :dot="true">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-    <nut-tabbar-item tab-title="标签" :value="110">
-      <template #icon>
-        <Home></Home>
-      </template>
-    </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-import { Home } from '@nutui/icons-vue';
-</script>
-```
-
-:::
+> demo: tabbar custom-quantity
 
 ### 固定底部，可自由跳转
 
-:::demo
-
-```vue
-<template>
-  <nut-tabbar bottom safe-area-inset-bottom placeholder>
-    <nut-tabbar-item v-for="(item, index) in List" :tab-title="item.title" :icon="item.icon"> </nut-tabbar-item>
-  </nut-tabbar>
-</template>
-<script setup>
-import { h } from 'vue';
-import { Home, Category, Find, Cart, My } from '@nutui/icons-vue';
-const List = [
-  {
-    title: '标签',
-    icon: h(Home)
-  },
-  {
-    title: '标签',
-    icon: h(Category)
-  },
-  {
-    title: '标签',
-    icon: h(Find)
-  },
-  {
-    title: '标签',
-    icon: h(Cart)
-  },
-  {
-    title: '标签',
-    icon: h(My)
-  }
-];
-</script>
-```
-
-:::
+> demo: tabbar fixed-bottom
 
 ## API
